LPS-93844 rename
